refactor: remove white spaces
