+++ pocket-ui/README.md
@@ -1,38 +1,34 @@
-
-
 ## Initial Screener Exercise
+
 This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-
-
 ### Features
+
 ---
 
-* Renders Pocket's logo centered at the top of the screen, and a Material UI button below the Pocket logo that reads `Annette Sunday`
-
-
+- Renders Pocket's logo centered at the top of the screen, and a Material UI button below the Pocket logo that reads `Annette Sunday`
 
 ### Technologies Used
+
 ---
 
 - Material UI
 - React
 - Typescript
 
-
 ### Installation
+
 ---
 
 - Clone this branch.
 - Change directory into the pocket-ui directory.
 - Install all necessary packages in the package.json file. Depending on if you are using `yarn` or `npm`, you can use the command below:
-> yarn install
-OR
-> npm install
+  > yarn install
+  > OR
+  > npm install
 - Run the command below to start the application locally:
-> yarn start
+  > yarn start
 
 > npm test -- --coverage
+
 - To run all the tests with coverage flag
-
-
