+++ pocket-ui/src/css/Pocket.css
@@ -1,6 +1,5 @@
 .Content-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
