+++ pocket-ui/src/index.tsx
@@ -1,12 +1,11 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './css/index.css';
-import App from './components/App';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./css/index.css";
+import App from "./components/App";
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
-
